--- a/RuntimeUnitTestToolkit/Assets/RuntimeUnitTestToolkit/RuntimeUnitTestToolkit.asmdef
+++ b/RuntimeUnitTestToolkit/Assets/RuntimeUnitTestToolkit/RuntimeUnitTestToolkit.asmdef
@@ -2,21 +2,18 @@
     "name": "RuntimeUnitTestToolkit",
     "rootNamespace": "",
     "references": [
-        "UnityEngine.TestRunner",
-        "UnityEditor.TestRunner",
         "Unity.PerformanceTesting"
+    ],
+    "optionalUnityReferences": [
+        "TestAssemblies"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
-    "overrideReferences": true,
-    "precompiledReferences": [
-        "nunit.framework.dll"
-    ],
-    "autoReferenced": false,
-    "defineConstraints": [
-        "UNITY_INCLUDE_TESTS"
-    ],
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
     "versionDefines": [],
     "noEngineReferences": false
 }