--- a/RuntimeUnitTestToolkit/Assets/RuntimeUnitTestToolkit/RuntimeUnitTestToolkit.asmdef
+++ b/RuntimeUnitTestToolkit/Assets/RuntimeUnitTestToolkit/RuntimeUnitTestToolkit.asmdef
@@ -1,6 +1,5 @@
 {
     "name": "RuntimeUnitTestToolkit",
-    "rootNamespace": "",
     "references": [
         "Unity.PerformanceTesting"
     ],