--- a/RuntimeUnitTestToolkit/Assets/RuntimeUnitTestToolkit/RuntimeUnitTestToolkit.asmdef
+++ b/RuntimeUnitTestToolkit/Assets/RuntimeUnitTestToolkit/RuntimeUnitTestToolkit.asmdef
@@ -13,6 +13,12 @@
     "precompiledReferences": [],
     "autoReferenced": true,
     "defineConstraints": [],
-    "versionDefines": [],
+    "versionDefines": [
+        {
+            "name": "com.unity.test-framework.performance",
+            "expression": "",
+            "define": "RUNTIMEUNITTESTTOOLKIT_PERFORMANCETESTING_SUPPORT"
+        }
+    ],
     "noEngineReferences": false
 }