--- a/RuntimeUnitTestToolkit/Assets/RuntimeUnitTestToolkit/UnitTestData.cs
+++ b/RuntimeUnitTestToolkit/Assets/RuntimeUnitTestToolkit/UnitTestData.cs
@@ -4,7 +4,9 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+#if RUNTIMEUNITTESTTOOLKIT_PERFORMANCETESTING_SUPPORT
 using Unity.PerformanceTesting;
+#endif
 using UnityEngine;
 using UnityEngine.TestTools;
 
@@ -59,6 +61,7 @@ namespace RuntimeUnitTestToolkit
 
                     foreach (var method in item.GetMethods())
                     {
+#if RUNTIMEUNITTESTTOOLKIT_PERFORMANCETESTING_SUPPORT
                         // Skip Test if PerformanceAttribute exists.
                         PerformanceAttribute t0 = null;
                         try
@@ -73,6 +76,7 @@ namespace RuntimeUnitTestToolkit
                         {
                             continue;
                         }
+#endif
 
                         TestAttribute t1 = null;
                         try