--- a/RuntimeUnitTestToolkit/Assets/Scripts/Tests/PerfTest.cs.meta
+++ b/RuntimeUnitTestToolkit/Assets/Scripts/Tests/PerfTest.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6a26f3ef6f940ab4eace21494a7b68b7
+timeCreated: 1488689733
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: