--- a/RuntimeUnitTestToolkit/Packages/manifest.json
+++ b/RuntimeUnitTestToolkit/Packages/manifest.json
@@ -4,6 +4,7 @@
     "com.unity.ide.visualstudio": "2.0.22",
     "com.unity.ide.vscode": "1.2.5",
     "com.unity.test-framework": "1.1.33",
+    "com.unity.test-framework.performance": "3.0.3",
     "com.unity.toolchain.win-x86_64-linux-x86_64": "2.0.9",
     "com.unity.ugui": "1.0.0"
   }