--- a/RuntimeUnitTestToolkit/Packages/manifest.json
+++ b/RuntimeUnitTestToolkit/Packages/manifest.json
@@ -7,5 +7,8 @@
     "com.unity.test-framework.performance": "3.0.3",
     "com.unity.toolchain.win-x86_64-linux-x86_64": "2.0.9",
     "com.unity.ugui": "1.0.0"
-  }
+  },
+  "testables": [
+    "com.unity.test-framework.performance"
+  ]
 }