## tl;dr;

RuntimeUnitTestToolkit failed to run test if method if performance testing. This PR skip test which has `Unity.PerformanceTesting.PerformanceAttribute`.

## Test Changes

<details><summary>Before (failed)</summary>

```
Begin Test Class: Group3
[OK]DivideTest(12, 3, 4), 0.02ms
[OK]DivideTest(12, 2, 6), 0.01ms
[OK]DivideTest(12, 4, 3), 0.01ms
[OK]DivideTest_2(12, 3, 4), 0.02ms
[OK]DivideTest_2(12, 2, 6), 0.02ms
[OK]DivideTest_2(12, 4, 3), 0.01ms
[OK]DivideTest2(12, 3, 4), 0.01ms
[OK]DivideTest2(12, 2, 6), 0.01ms
[OK]DivideTest2(12, 4, 3), 0.01ms
[OK]DivideTest5(12, 3), 0.00ms
[OK]DivideTest5(12, 2), 0.00ms
[OK]DivideTest5(12, 4), 0.00ms
[OK]DivideTest4(12, 3), 0.00ms
[OK]DivideTest4(12, 2), 0.00ms
[OK]DivideTest4(12, 4), 0.00ms
[OK]DivideTest3(12, 3, 4), 0.02ms
[OK]DivideTest3(12, 2, 6), 0.01ms
[OK]DivideTest3(12, 4, 3), 0.01ms
Begin Test Class: PerfTest
[NG]DummySerialize, Unity.PerformanceTesting.Exceptions.PerformanceTestException: Trying to record samples but there is no active performance tests.
  at Unity.PerformanceTesting.PerformanceTest.GetSampleGroup (System.String name) [0x00000] in <00000000000000000000000000000000>:0
  at Unity.PerformanceTesting.Measure.Custom (Unity.PerformanceTesting.SampleGroup sampleGroup, System.Double value) [0x00000] in <00000000000000000000000000000000>:0
  at Unity.PerformanceTesting.Measurements.MethodMeasurement.Run () [0x00000] in <00000000000000000000000000000000>:0
  at RuntimeUnitTestToolkit.UnitTestRunner+<RunTestInCoroutine>d__16.MoveNext () [0x00000] in <00000000000000000000000000000000>:0
  at UnityEngine.SetupCoroutine.InvokeMoveNext (System.Collections.IEnumerator enumerator, System.IntPtr returnValueAddress) [0x00000] in <00000000000000000000000000000000>:0
[NG]DummyDeserialize, Unity.PerformanceTesting.Exceptions.PerformanceTestException: Trying to record samples but there is no active performance tests.
  at Unity.PerformanceTesting.PerformanceTest.GetSampleGroup (System.String name) [0x00000] in <00000000000000000000000000000000>:0
  at Unity.PerformanceTesting.Measure.Custom (Unity.PerformanceTesting.SampleGroup sampleGroup, System.Double value) [0x00000] in <00000000000000000000000000000000>:0
  at Unity.PerformanceTesting.Measurements.MethodMeasurement.Run () [0x00000] in <00000000000000000000000000000000>:0
  at RuntimeUnitTestToolkit.UnitTestRunner+<RunTestInCoroutine>d__16.MoveNext () [0x00000] in <00000000000000000000000000000000>:0
  at UnityEngine.SetupCoroutine.InvokeMoveNext (System.Collections.IEnumerator enumerator, System.IntPtr returnValueAddress) [0x00000] in <00000000000000000000000000000000>:0
Begin Test Class: SampleGroup
[Log]Called OneTimeSetUp
Called OneTimeSetUp
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:107
Unity Setup One:107
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:108
Unity Setup Two:108
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[OK]SumTest, 0.02ms
[Log]Called Teardown
Called Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:<>m__Finally2()
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown One:108
Unity Teardown One:108
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown Two:109
Unity Teardown Two:109
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:110
Unity Setup One:110
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:111
Unity Setup Two:111
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[OK]ParameterTest(1, 2, 3), 0.03ms
[Log]Called Teardown
Called Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:<>m__Finally2()
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown One:111
Unity Teardown One:111
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown Two:112
Unity Teardown Two:112
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:113
Unity Setup One:113
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:114
Unity Setup Two:114
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[OK]ParameterTest(10, 20, 30), 0.02ms
[Log]Called Teardown
Called Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:<>m__Finally2()
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown One:114
Unity Teardown One:114
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown Two:115
Unity Teardown Two:115
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:116
Unity Setup One:116
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:117
Unity Setup Two:117
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[OK]ParameterTest(100, 200, 300), 0.03ms
[Log]Called Teardown
Called Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:<>m__Finally2()
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown One:117
Unity Teardown One:117
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown Two:118
Unity Teardown Two:118
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:119
Unity Setup One:119
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:120
Unity Setup Two:120
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[OK]AsyncTest, 0.86ms
[Log]Called Teardown
Called Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:<>m__Finally2()
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown One:180
Unity Teardown One:180
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown Two:181
Unity Teardown Two:181
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called OneTimeTearDown
Called OneTimeTearDown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

Begin Test Class: TestCaseSourceGroup
[OK]DivideTest(12, 3, 4), 0.03ms
[OK]DivideTest(12, 2, 6), 0.02ms
[OK]DivideTest(12, 4, 3), 0.01ms
[OK]DivideUnityTest(12, 3, 4), 0.07ms
[OK]DivideUnityTest(12, 2, 6), 0.05ms
[OK]DivideUnityTest(12, 4, 3), 0.05ms
[Log]Global Teardown
Global Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<ExecuteAllInCoroutine>d__17:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

Test Failed, please see [NG] log.
```
</details> 

<details><summary>After (success)</summary>

```
[PhysX] Initialized MultithreadedTaskDispatcher with 16 workers.
UnloadTime: 0.623400 ms
[Log][SKIP]DummySerialize has PerformanceAttribute.
[SKIP]DummySerialize has PerformanceAttribute.
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<GetTestTargetTypes>d__9:MoveNext()
RuntimeUnitTestToolkit.UnitTestData:.ctor()
RuntimeUnitTestToolkit.UnitTestRunner:Start()

[Log][SKIP]DummyDeserialize has PerformanceAttribute.
[SKIP]DummyDeserialize has PerformanceAttribute.
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<GetTestTargetTypes>d__9:MoveNext()
RuntimeUnitTestToolkit.UnitTestData:.ctor()
RuntimeUnitTestToolkit.UnitTestRunner:Start()

ERROR: Shader UI/Default shader is not supported on this GPU (none of subshaders/fallbacks are suitable)
[Log]Global Setup
Global Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<ExecuteAllInCoroutine>d__17:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.UnitTestRunner:Start()

Begin Test Class: Group2
[Log]Called OneTimeSetUp
Called OneTimeSetUp
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:3
Unity Setup One:3
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:4
Unity Setup Two:4
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[OK]SumTest, 1.07ms
[Log]Called Teardown
Called Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:<>m__Finally2()
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown One:4
Unity Teardown One:4
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown Two:5
Unity Teardown Two:5
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:6
Unity Setup One:6
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:7
Unity Setup Two:7
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[OK]ParameterTest(1, 2, 3), 0.03ms
[Log]Called Teardown
Called Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:<>m__Finally2()
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown One:7
Unity Teardown One:7
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown Two:8
Unity Teardown Two:8
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:9
Unity Setup One:9
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:10
Unity Setup Two:10
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[OK]ParameterTest(10, 20, 30), 0.02ms
[Log]Called Teardown
Called Teardown
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:<>m__Finally2()
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown One:10
Unity Teardown One:10
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Teardown Two:11
Unity Teardown Two:11
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnityTearDown>d__9:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Called Setup
Called Setup
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup One:12
Unity Setup One:12
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)
RuntimeUnitTestToolkit.<RunTestInCoroutine>d__16:MoveNext()
UnityEngine.SetupCoroutine:InvokeMoveNext(IEnumerator, IntPtr)

[Log]Unity Setup Two:13
Unity Setup Two:13
UnityEngine.Logger:Log(LogType, Object)
UnityEngine.Debug:Log(Object)
SampleUnitTest.<UnitySetupRun>d__8:MoveNext()
…